requirements.txt
fastapi>=0.115.8,<0.116.0
python-multipart>=0.0.20,<0.0.21
python-dotenv>=1.0.1,<2.0.0
uvicorn>=0.34.0,<0.35.0
pydantic==2.10.4
pydantic-core==2.27.2
toml==0.10.2
loguru>=0.6.0,<0.7.0
starlette>=0.45.3,<0.46.0
# sqlalchemy>=2.0.38,<3.0.0
aioboto3>=13.4.0,<14.0.0
mysql-connector-python>=9.2.0,<10.0.0
mysqlclient>=2.2.7,<3.0.0
aiomysql>=0.2.0,<0.3.0
cryptography>=44.0.0,<45.0.0
numpy>=2.2.5,<3.0.0
langchain>=0.3.23,<0.4.0
langchain-community>=0.3.21,<0.4.0
langchain-core>=0.3.55,<0.4.0
unstructured>=0.17.2,<0.18.0
langchain-openai>=0.3.14,<0.4.0
python-docx>=1.1.2,<2.0.0
pymilvus==2.5.3
